Backend/requirements.txt
# Django core
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg[binary]>=3.1.0

# CORS
django-cors-headers>=4.3.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client for OpenRouter API
httpx>=0.27.0

# Development
django-extensions>=3.2.0

# Payment Integration
paypalrestsdk==1.13.3
